% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.r
\name{.plot_hist_density}
\alias{.plot_hist_density}
\title{Plot Histogram With Optional Density Overlay}
\usage{
.plot_hist_density(
  x_vec,
  main_text = NULL,
  label = NULL,
  is_diag = FALSE,
  main = TRUE,
  plot_args = list(),
  ...
)
}
\arguments{
\item{x_vec}{Numeric vector to plot.}

\item{label}{Character; label for both x- and y-axis (or diagonal panels).}

\item{is_diag}{Logical: whether plotting in a diagonal panel within pairs plot}

\item{main}{Logical; if TRUE, display the main title specified in \code{main_text}.}

\item{plot_args}{List of graphical  parameters (colors, breaks, line widths, etc.)}

\item{...}{Additional graphical parameters passed to \code{hist()} or \code{lines()}}
}
\description{
Internal helper for plotting histograms with an optional kernel density curve.
For diagonal panels in pairs plots, counts and density are scaled to \link{0,1}.

Skips plotting if data has only one unique value.
}
\keyword{internal}
