% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.r
\name{printSimData}
\alias{printSimData}
\alias{print.LMMdata}
\alias{print.GLMMdata}
\alias{print.LMdata}
\title{Print Simulated Dataset}
\usage{
\method{print}{LMMdata}(object, replication = 1, iteration = 1)

\method{print}{GLMMdata}(object, replication = 1, iteration = 1)

\method{print}{LMdata}(object, replication = 1, iteration = 1)

\method{print}{GLMMdata}(object, replication = 1, iteration = 1)
}
\arguments{
\item{object}{Numeric vector, matrix, named list, or simulation object
(of class \code{LMMdata}, \code{GLMMdata}, \code{LMdata}, \code{GLMdata}, \code{IIDdata}).
If a list, it must include the components required for the selected \code{type}.}

\item{replication}{Integer; which replication (dataset) to plot
(LMM/GLMM, LM/GLM only; default \code{1}).}

\item{iteration}{Integer; which iteration within the replication to plot
(LMM/GLMM, LM/GLM only; default \code{1}).}
}
\description{
Prints a dataset from a simulation object, showing metadata and actual data.
Supports multiple types of simulated datasets:
\itemize{
\item \strong{Mixed Models} (of class \code{LMMdata} or \code{GLMMdata})
\item \strong{Regression Models} (of class \code{LMdata} or \code{GLMdata})
\item \strong{IID samples} (of class \code{IIDdata})
}

The function also works for named lists containing:
\itemize{
\item response vectors as \code{y}
\item fixed-effects design matrix as \code{X}
\item optional random-effects design matrix as \code{Z}
\item optional random-effects realizations as \code{RE}
}

Numeric vectors or matrices are treated as IID samples (columns = variables).
}
\details{
The function prints:
\itemize{
\item Metadata including number of subjects/observations, dimensions of design matrices,
regression or fixed-effects coefficients, \code{SNR}, and \code{sigma_e}.
\item Optional distribution names if \code{distr_settings} is provided within object.
\item The actual dataset (\code{y}, \code{X}, \code{Z}, \code{RE}) for the specified replication and iteration.
}
}
\examples{
\dontrun{
# Mixed model
sim_data <- simulate_LMMdata(c(2,3), sim_spec(p=5))
print(sim_data, replication = 1, iteration = 2)

# Regression model
sim_data_lm <- simulate_LMdata(c(2,3), sim_spec(p=5))
print(sim_data_lm)

# IID samples
iid_data <- matrix(rnorm(50), ncol=5)
print(iid_data)

# Named list
mylist <- list(y = rnorm(10), X = matrix(rnorm(20), ncol=2))
print(mylist)
}

}
